refactor: :recycle: Move tests into separate files
